2018/06/25

+++ webday17/src/com/itheima/product/web/servlet/ProductDeleteAllServlet.java
@@ -17,8 +17,10 @@ public class ProductDeleteAllServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		
 		String[] ids = request.getParameterValues("ids");
-		ProductService productService = new ProductService();
-		productService.deleteAll(ids);
+		if(ids != null) {
+			ProductService productService = new ProductService();
+			productService.deleteAll(ids);
+		}
 		request.getRequestDispatcher("/ProductServlet").forward(request, response);
 	}
